pyproject.toml
[project]
name = "py-fast-transaction-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.134.0",
    "python-dotenv>=1.2.1",
    "snowflake-id>=1.0.2",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"